Add ssh-agent support for Windows
Add support for using ssh-agent on Windows. The Win32 implementation of
ssh-agent uses a named pipe rather than a unix domain socket on the
listener but the underlying protocol is still the same.
