+++ src/Tmds.Ssh/SshAgent.cs
@@ -4,7 +4,7 @@
 using System.Buffers;
 using System.Buffers.Binary;
 using System.Diagnostics;
-using System.Net;
+using System.IO.Pipes;
 using System.Net.Sockets;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -23,42 +23,28 @@ namespace Tmds.Ssh
         private const uint SSH_AGENT_RSA_SHA2_512 = 4;
         private const int MaxPacketSize = int.MaxValue; // We trust the SSH Agent.
 
-        private static EndPoint? _defaultEndPoint;
-
         public struct Identity
         {
             public string Comment { init; get; }
             public SshKey PublicKey { init; get; }
         }
 
-        public static EndPoint? DefaultEndPoint
+        public static string? DefaultEndPoint
         {
             get
             {
-                string? sshAuthSock = Environment.GetEnvironmentVariable("SSH_AUTH_SOCK");
-
-                if (OperatingSystem.IsWindows())
-                {
-                    return null;
-                }
-
-                if (string.IsNullOrEmpty(sshAuthSock))
-                {
-                    return null;
-                }
-
-                _defaultEndPoint ??= new UnixDomainSocketEndPoint(sshAuthSock);
-
-                return _defaultEndPoint;
+                return OperatingSystem.IsWindows()
+                    ? "openssh-ssh-agent"
+                    : Environment.GetEnvironmentVariable("SSH_AUTH_SOCK");
             }
         }
 
-        private readonly EndPoint _endPoint;
+        private readonly string _endPoint;
         private readonly SequencePool _sequencePool;
 
-        private SocketSshConnection? _agentConnection;
+        private StreamSshConnection? _agentConnection;
 
-        public SshAgent(EndPoint endPoint, SequencePool sequencePool)
+        public SshAgent(string endPoint, SequencePool sequencePool)
         {
             _endPoint = endPoint;
             _sequencePool = sequencePool;
@@ -66,17 +52,48 @@ namespace Tmds.Ssh
 
         public async ValueTask<bool> TryConnect(CancellationToken cancellationToken)
         {
+            NamedPipeClientStream? pipe = null;
             Socket? socket = null;
             try
             {
-                socket = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
-                await socket.ConnectAsync(_endPoint, cancellationToken).ConfigureAwait(false);
-                _agentConnection = new SocketSshConnection(NullLoggerFactory.Instance.CreateLogger<SshClient>(), _sequencePool, socket);
+                var logger = NullLoggerFactory.Instance.CreateLogger<SshClient>();
+
+                if (OperatingSystem.IsWindows())
+                {
+                    // There is no easy way to detect if a named pipe exists or
+                    // not. File.Exists is the fastest but will actually open a
+                    // connection and close it. Using this won't open the pipe
+                    // Directory.GetFiles("\\\\.\\pipe\\", endPoint).Length > 0
+                    // But it is slower.
+                    if (!File.Exists(@$"\\.\pipe\{_endPoint}"))
+                    {
+                        return false;
+                    }
+
+                    pipe = new NamedPipeClientStream(
+                        ".",
+                        _endPoint,
+                        PipeDirection.InOut,
+                        PipeOptions.Asynchronous,
+                        System.Security.Principal.TokenImpersonationLevel.Anonymous);
+                    await pipe.ConnectAsync(cancellationToken).ConfigureAwait(false);
+
+                    _agentConnection = new StreamSshConnection(logger, _sequencePool, pipe);
+                }
+                else
+                {
+                    socket = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
+                    await socket.ConnectAsync(new UnixDomainSocketEndPoint(_endPoint), cancellationToken).ConfigureAwait(false);
+
+                    _agentConnection = new SocketSshConnection(logger, _sequencePool, socket);
+                }
+
                 _agentConnection.SetEncryptorDecryptor(new SshAgentPacketEncryptor(), new SshAgentPacketDecryptor(_sequencePool), false, false);
                 return true;
             }
             catch
             {
+                pipe?.Dispose();
                 socket?.Dispose();
                 return false;
             }
@@ -101,7 +118,7 @@ namespace Tmds.Ssh
             _agentConnection?.Dispose();
         }
 
-        private SocketSshConnection GetAgentConnection()
+        private StreamSshConnection GetAgentConnection()
         {
             return _agentConnection ?? throw new InvalidOperationException("Not connected");
         }
