+++ src/Tmds.Ssh/SshSession.cs
@@ -278,7 +278,7 @@ sealed partial class SshSession
 
         await HandleConnectionAsync(connection, ConnectionInfo).ConfigureAwait(false);
     }
-    
+
     private async Task HandleConnectionAsync(SshConnection connection, SshConnectionInfo ConnectionInfo)
     {
         Task sendTask = SendLoopAsync(connection);
