+++ test/Tmds.Ssh.Tests/TestServer.cs
@@ -41,7 +41,7 @@ sealed class TestServer : IAsyncDisposable
     {
         ILogger<SshClient> logger = new NullLoggerFactory().CreateLogger<SshClient>();
         Socket clientSocket = await _serverSocket.AcceptAsync().ConfigureAwait(false);
-        using SocketSshConnection socketSshConnection = new SocketSshConnection(logger, new SequencePool(), clientSocket);
+        using StreamSshConnection socketSshConnection = new SocketSshConnection(logger, new SequencePool(), clientSocket);
         await _handler(socketSshConnection);
     }
